Add option to allocate empty group on heap
We were using phmap with an allocator backed by a memory mapped file.
When we loaded the file again and mapped it back into memory, we got
intermittent crashes. The culprit was `EmptyGroup()`.

`EmptyGroup()` was allocating `empty_group` inside the process space,
and not inside the memory mapped space. When another process loaded up
the memory mapped file, the `empty_group` pointed outside the address
space. This caused a crash.

To fix this, `PHMAP_NO_STATIC_EMPTY_GROUP` is introduced. If it's
defined, phmap will use the allocator given to it to allocate the
`empty_group`.